backend/requirements.txt
# Core Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pydantic>=2.7.4,<3.0.0
python-dotenv>=1.0.1

# Database
sqlalchemy>=2.0.30
asyncpg>=0.29.0
alembic>=1.13.1
databases>=0.8.0
psycopg2-binary>=2.9.9
pgvector>=0.3.2

# Authentication & Security
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
PyJWT>=2.8.0

# AI/ML & LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-google-genai>=2.0.0
langchain-anthropic>=0.3.0
langchain-chroma>=0.1.4
anthropic>=0.39.0
faiss-cpu==1.11.0.post1
numpy>=1.26.4
scikit-learn>=1.5.0

# Google Cloud Services
google-genai>=1.0.0
google-cloud-talent>=2.17.2
google-cloud-texttospeech>=2.16.1
google-cloud-speech>=2.27.0

# Document Processing
pdf2image>=1.17.0
pypdf>=4.0.0
python-docx>=1.2.0
reportlab>=4.1.0
Pillow>=10.3.0  # Added - required for image processing

# Web Scraping & HTTP
httpx>=0.27.0
aiohttp>=3.9.5
websockets>=12.0
requests>=2.32.2
playwright>=1.52.0
beautifulsoup4>=4.12.3
lxml>=5.2.1
readability-lxml>=0.8.1

# Email & Communication
fastapi-mail>=1.4.1

# Payment Processing
stripe>=10.0.0

# Monitoring & Observability
opentelemetry-api>=1.24.0
opentelemetry-sdk>=1.24.0
opentelemetry-instrumentation>=0.45b0
opentelemetry-instrumentation-fastapi>=0.45b0
opentelemetry-instrumentation-sqlalchemy>=0.45b0
opentelemetry-exporter-otlp>=1.24.0

# Development & Testing
pytest>=8.2.0
pytest-asyncio>=0.23.0